Semana 6 - Rodrigo Herman
Semana 6 - Rodrigo Herman

+++ src/usuario/usuario.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, ConflictException, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { ListaUsuarioDTO } from './dto/ListaUsuario.dto';
 import { UsuarioEntity } from './usuario.entity';
@@ -13,7 +13,26 @@ export class UsuarioService {
   ) {}
 
   async criaUsuario(usuarioEntity: UsuarioEntity) {
-    await this.usuarioRepository.save(usuarioEntity);
+    try{
+      const usuario = await this.buscaPorEmail(usuarioEntity.email);
+
+      if (usuario !== null)
+      {
+        throw new ConflictException("Usuário já existe para este e-mail!");
+      }
+    }
+    catch(erro)
+    {
+      if (erro instanceof NotFoundException){
+        // Se não encontrou o usuário, procede com a inclusão
+        await this.usuarioRepository.save(usuarioEntity);
+      }
+      else
+      {
+        // Se for outra excecao, joga o erro adiante;
+        throw erro;
+      }
+    }    
   }
 
   async listUsuarios() {
@@ -28,14 +47,35 @@ export class UsuarioService {
     const checkEmail = await this.usuarioRepository.findOne({
       where: { email },
     });
+
+    if (checkEmail === null)
+    {
+      throw new NotFoundException('Usuário não encontrado com este e-mail!');
+    }
+
     return checkEmail;
   }
 
   async atualizaUsuario(id: string, novosDados: AtualizaUsuarioDTO) {
+    const usuario = await this.usuarioRepository.findOne({ where: {id} });
+
+    if (usuario === null)
+    {
+      throw new BadRequestException('Usuário informado não existe!');
+    }
+
     await this.usuarioRepository.update(id, novosDados);
   }
 
   async deletaUsuario(id: string) {
+
+    const usuario = await this.usuarioRepository.findOne({ where: {id} });
+
+    if (usuario === null)
+    {
+      throw new BadRequestException('Usuário informado não existe!');
+    }
+
     await this.usuarioRepository.delete(id);
   }
 }
